[refactor/#87]: long으로 변경
 - 레파지토리에서 형변환하지말고 서비스에서 변경할 수 잇도록

+++ src/main/java/com/woohaengshi/backend/repository/statistics/StatisticsCustomRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.domain.Pageable;
 import java.util.List;
 
 public interface StatisticsCustomRepository {
-    public int getMemberRank(StatisticsType statisticsType, Statistics statistics);
+    public long getMemberRank(StatisticsType statisticsType, Statistics statistics);
 
     public List<Statistics> findStatisticsByTypeSortedByTimeDesc(
             StatisticsType statisticsType, Pageable pageable);
